+++ integration/cursors/tailable_test.go
@@ -323,53 +323,3 @@ func TestCursorsTailableTwoCursorsSameCollection(t *testing.T) {
 		assert.Equal(t, cursorID2, nextID2)
 	})
 }
-
-// getFirstBatch takes the response from the query that generates the cursors,
-// validates if it contains cursor.firstBatch, and cursor ID, and returns those.
-func getFirstBatch(t testing.TB, res bson.D) (*types.Array, any) {
-	t.Helper()
-
-	doc := integration.ConvertDocument(t, res)
-
-	v, _ := doc.Get("cursor")
-	require.NotNil(t, v)
-
-	cursor, ok := v.(*types.Document)
-	require.True(t, ok)
-
-	cursorID, _ := cursor.Get("id")
-	assert.NotNil(t, cursorID)
-
-	v, _ = cursor.Get("firstBatch")
-	require.NotNil(t, v)
-
-	firstBatch, ok := v.(*types.Array)
-	require.True(t, ok)
-
-	return firstBatch, cursorID
-}
-
-// getNextBatch takes the response from the getMore query,
-// validates if it contains cursor.nextBatch, and cursor ID, and returns those.
-func getNextBatch(t testing.TB, res bson.D) (*types.Array, any) {
-	t.Helper()
-
-	doc := integration.ConvertDocument(t, res)
-
-	v, _ := doc.Get("cursor")
-	require.NotNil(t, v)
-
-	cursor, ok := v.(*types.Document)
-	require.True(t, ok)
-
-	cursorID, _ := cursor.Get("id")
-	assert.NotNil(t, cursorID)
-
-	v, _ = cursor.Get("nextBatch")
-	require.NotNil(t, v)
-
-	firstBatch, ok := v.(*types.Array)
-	require.True(t, ok)
-
-	return firstBatch, cursorID
-}
